adicionado lombok na classe de graduacao
