+++ src/main/java/com/br/ifspapi/anexo/AnexoDTO.java
@@ -1,15 +1,18 @@
 package com.br.ifspapi.anexo;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.convention.MatchingStrategies;
 
+@Getter @Setter
+@AllArgsConstructor
+@NoArgsConstructor
 public class AnexoDTO {
 
-    @Getter @Setter
     private Long id;
-    @Getter @Setter
     private String src;
 
     public static AnexoDTO from(AnexoModel anexoModel){
