+++ src/main/java/com/br/ifspapi/anexo/AnexoModel.java
@@ -1,9 +1,6 @@
 package com.br.ifspapi.anexo;
 
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.convention.MatchingStrategies;
 import org.modelmapper.spi.MatchingStrategy;
@@ -12,15 +9,15 @@ import javax.persistence.*;
 
 @Entity
 @EqualsAndHashCode
+@AllArgsConstructor
 @NoArgsConstructor
 @Table(name = "anexo")
+@Getter @Setter
 public class AnexoModel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Getter @Setter
     private Long id;
-    @Getter @Setter
     private String src;
 
     public AnexoModel(String src) {
