+++ src/main/java/com/br/ifspapi/graduation/GraduationDto.java
@@ -1,4 +1,4 @@
-package com.br.ifspapi.anexo;
+package com.br.ifspapi.graduation;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,17 +7,21 @@ import lombok.Setter;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.convention.MatchingStrategies;
 
-@Getter @Setter
-@NoArgsConstructor
+
+@Getter
+@Setter
 @AllArgsConstructor
-public class AnexoForm {
+@NoArgsConstructor
+public class GraduationDto {
 
-    private String src;
+    private Long id;
+    private String name;
+    private String descricao;
 
-    public static AnexoForm from(AnexoModel anexoModel){
+    public static GraduationDto from(Graduation graduation) {
         ModelMapper modelMapper = new ModelMapper();
         modelMapper.getConfiguration().setMatchingStrategy(MatchingStrategies.STRICT);
-        return modelMapper.map(anexoModel, AnexoForm.class);
+        return modelMapper.map(graduation, GraduationDto.class);
     }
 
 }
