+++ src/test/java/com/br/ifspapi/IfspApiApplicationTests.java
@@ -1,15 +1,8 @@
 package com.br.ifspapi;
 
-import com.br.ifspapi.graduation.dto.GraduationForm;
-import com.br.ifspapi.graduation.model.Graduation;
-import com.br.ifspapi.graduation.repository.GraduationRepository;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.Rollback;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 
 @SpringBootTest
